requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
pymongo
python-dotenv
pydub
